% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scprep_biomart.R
\name{scprep_biomart}
\alias{scprep_biomart}
\title{Function to annotate feature data in ExpressionSet, Seurat, or SingleCellExperiment objects}
\usage{
scprep_biomart(
  dataset = dataset,
  ensembl_target = ensembl_target,
  reference = reference
)
}
\arguments{
\item{dataset}{ExpressionSet, Seurat, or SingleCellExperiment object}

\item{ensembl_target}{Species input for biomaRt feature annotation}

\item{reference}{Reference genome (on annotation.csv & SampleSheet.csv): "refdata-cellranger-GRCh38-3.0.0" or "GRCh38_pre-mRNA" for Ensembl 93; "refdata-gex-GRCh38-2020-A" or "GRCh38_pre-mRNA-2020-A" for Ensembl 98}
}
\value{
Object of same type as input containing annotated feature data
}
\description{
This function annotates feature data with scRNA data using biomaRt
}
\examples{
\dontrun{
scprep_biomart(dataset=object, ensembl_target="hsapiens_gene_ensembl", 
               reference="refdata-cellranger-GRCh38-3.0.0")
}
}
