% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scprep_cell_filter_multi.R
\name{scprep_cell_filter_multi}
\alias{scprep_cell_filter_multi}
\title{Function for multi-sample discrimination of "Cell" barcodes from "Dead" cells and "Debris" based on hard cut-offs for UMI transcript & gene content, and percent mitochondrial UMIs}
\usage{
scprep_cell_filter_multi(
  dataset = dataset,
  min_umi = 100,
  min_gene = 0,
  max_mito = 0.2
)
}
\arguments{
\item{dataset}{ExpressionSet, Seurat, or SingleCellExperiment object with scRNA data}

\item{min_umi}{UMI transcript threshold}

\item{min_gene}{Genes detected threshold}

\item{max_mito}{Percent mitochondrial transcript threshold (0 = No mito transcripts allowed, 1 = No limit for mito transcripts)}
}
\value{
Multi-sample Cell, Dead, & Debris assignment
}
\description{
Across all samples (Sample designation in object metadata slot), this function designates high transcript/gene content & low percent mitochondrial content barcodes as "Cell", high percent mitochondrial content barcodes as "Dead", & low transcript/gene content barcodes as "Debris"
}
\examples{
scprep_cell_filter_multi(dataset=object, min_umi=100, min_gene=0, max_mito=0.2)
}
