% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scprep_eset_biomart.R
\name{scprep_eset_biomart}
\alias{scprep_eset_biomart}
\title{Function to annotate feature data (fData) in ExpressionSet object}
\usage{
scprep_eset_biomart(
  dataset = dataset,
  ensembl_target = ensembl_target,
  reference = reference
)
}
\arguments{
\item{dataset}{Existing ExpressionSet object created with eset_build()}

\item{ensembl_target}{Species input for biomaRt feature annotation}

\item{reference}{Reference genome (on annotation.csv & SampleSheet.csv): "refdata-cellranger-GRCh38-3.0.0" or "GRCh38_pre-mRNA" for Ensembl 93; "refdata-gex-GRCh38-2020-A" or "GRCh38_pre-mRNA-2020-A" for Ensembl 98}
}
\value{
ExpressionSet object containing annotated feature data in fData slot
}
\description{
This function annotate feature data (fData) in ExpressionSet object with scRNA data
}
\examples{
scprep_eset_biomart(dataset=ExpressionSet, ensembl_target="hsapiens_gene_ensembl", reference="refdata-cellranger-GRCh38-3.0.0")
}
