% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scprep_seeds.R
\name{scprep_seeds}
\alias{scprep_seeds}
\title{Random seed generator for analysis reproducibility}
\usage{
scprep_seeds(n.seeds = n.seeds)
}
\arguments{
\item{n.seeds}{Number of random seeds}
}
\value{
Random seeds
}
\description{
This function generates 8-digit random seeds
}
\examples{
scprep_seeds(n.seeds=1000)
}
