% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template_scprep.R
\name{template_scprep}
\alias{template_scprep}
\title{Template function for building scRNA-Seq ExpressionSet object}
\usage{
template_scprep(dir_output = dir_output)
}
\arguments{
\item{dir_output}{Output directory for analysis results}
}
\value{
ExpressionSet object
}
\description{
This function builds an ExpressionSet object with scRNA data
}
\examples{
template_scprep(dir_output="/path/to/output")
}
