% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scprep_eset_build.R
\name{scprep_eset_build}
\alias{scprep_eset_build}
\title{Function to initiate single-cell data object}
\usage{
scprep_eset_build(
  sample_paths = sample_paths,
  annotation = annotation,
  file_type = "h5",
  gene_id = "symbol",
  vdj = vdj,
  cite = cite,
  atac = atac
)
}
\arguments{
\item{sample_paths}{Character vector of paths to filtered_feature_bc_matrix.h5 for all samples in project}

\item{annotation}{Matrix of project-specific annotation.csv}

\item{vdj}{Data is multimodal RNA & VDJ}

\item{cite}{Data is multimodal RNA & Protein}

\item{cite_ignore}{FALSE to ignore; Default is TRUE, ignore the "Antibody Capture" slot in the h5 file input and only extract/store the "Gene Expression" slot; This is only relevant if your data contains both RNA and protein modalities and you want to ignore the protein modality}

\item{atac}{Data is multimodal RNA & ATAC}
}
\value{
ExpressionSet, Seurat, or SingleCellExperiment object containing expression data for all samples & annotation
}
\description{
This function builds an ExpressionSet, Seurat, or SingleCellExperiment object with scRNA data
}
\examples{
scprep_eset_build(sample_paths=sample_paths, annotation=annotation, file_type="h5", vdj=vdj, cite=cite, atac=atac)
}
