% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template_scprep.R
\name{template_scprep}
\alias{template_scprep}
\title{Template function for building scRNA-Seq data objects}
\usage{
template_scprep(dir_output = dir_output)
}
\arguments{
\item{dir_output}{Output directory for analysis results}
}
\value{
ExpressionSet, Seurat, or SingleCellExperiment object (depending on output_type parameter)
}
\description{
This function builds an ExpressionSet, Seurat, or SingleCellExperiment object with scRNA data
}
\examples{
# Build ExpressionSet (default)
template_scprep(dir_output="/path/to/output")
# Output type can be specified in scprep_parameters.csv
}
